[+]se agrega alertas para cada formulario/condicionales para estas mismas alertas

+++ frontend/src/app/pages/private/categories/category-list/category-list.ts
@@ -1,7 +1,8 @@
 import { Component } from '@angular/core';
 import { AsyncPipe } from '@angular/common';
-import { BehaviorSubject, Observable, switchMap, Subscription } from 'rxjs';
 import { Router, RouterLink } from '@angular/router';
+import { BehaviorSubject, Observable, switchMap, Subscription } from 'rxjs';
+import Swal from 'sweetalert2';
 
 import { HttpCategories } from '../../../../core/service/http-categories';
 
@@ -31,15 +32,37 @@ export class CategoryList {
   }
 
   deleteCategory(_id: string): void {
-    this.subscription = this.httpCategories.deleteCategory(_id).subscribe({
-      next: (response) => {
-        console.log(response);
-        this.refreshCategoriesTrigger$.next();
-      },
-      error: (error) => {
-        console.error('Error deleting category:', error);
+
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+
+        this.subscription = this.httpCategories.deleteCategory(_id).subscribe({
+        next: (response) => {
+          console.log(response);
+          this.refreshCategoriesTrigger$.next();
+        },
+        error: (error) => {
+          console.error('Error deleting category:', error);
+        }
+    });
+
+    Swal.fire({
+      title: "Deleted!",
+      text: "Your file has been deleted.",
+      icon: "success"
+    });
       }
     });
+
+    
   }
 
   onEdit(_id: string): void {
