+++ frontend/src/app/pages/private/products/products-list/products-list.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 import { Router, RouterLink } from '@angular/router';
 import { AsyncPipe, CurrencyPipe } from '@angular/common';
 import { BehaviorSubject, Observable, Subscription, switchMap } from 'rxjs';
+import Swal from 'sweetalert2';
 
 import { HttpProducts } from '../../../../core/service/http-products';
 
@@ -32,15 +33,36 @@ export class ProductsList {
   }
 
   deleteProduct(_id: string): void {
-    this.subscription = this.httpProducts.deleteProduct(_id).subscribe({
-      next: (response) => {
-        console.log(response);
-        this.refreshProductsTrigger$.next();
-      },
-      error: (error) => {
-        console.error('Error deleting product:', error);
+
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+
+          this.subscription = this.httpProducts.deleteProduct(_id).subscribe({
+          next: (response) => {
+            console.log(response);
+            this.refreshProductsTrigger$.next();
+          },
+          error: (error) => {
+            console.error('Error deleting product:', error);
+          }
+    });
+
+        Swal.fire({
+          title: "Deleted!",
+          text: "Your file has been deleted.",
+          icon: "success"
+        });
       }
     });
+
   }
 
   ngOnDestroy(): void {
